- Добавил абстрактный класс - Добавил похожим классам наследование от общего
