+++ src/view/load-more-button-view.js
@@ -1,23 +1,9 @@
-import {createElement} from '../render';
+import AbstractView from './abstract-view';
 
 const createLoadMoreButtonTemplate = () => '<button class="films-list__show-more">Show more</button>';
 
-export default class LoadMoreButtonView {
-  #element = null;
-
-  get element () {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class LoadMoreButtonView extends AbstractView {
   get template() {
     return createLoadMoreButtonTemplate();
-  }
-
-  removeElement = () => {
-    this.#element = null;
   }
 }
