+++ src/view/movie-list.view.js
@@ -1,23 +1,9 @@
-import {createElement} from '../render';
+import AbstractView from './abstract-view';
 
 const createMovieListTemplate = () => '<div class="films-list__container"></div>';
 
-export default class MovieListView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class MovieListView extends AbstractView {
   get template() {
     return createMovieListTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
