+++ src/view/user-rank-view.js
@@ -1,4 +1,4 @@
-import {createElement} from '../render';
+import AbstractView from './abstract-view';
 
 const createUserRankTemplate = () => (
   `<section class="header__profile profile">
@@ -7,22 +7,8 @@ const createUserRankTemplate = () => (
   </section>`
 );
 
-export default class UserRankView {
-  #element = null;
-
-  get element () {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class UserRankView extends AbstractView {
   get template() {
     return createUserRankTemplate();
-  }
-
-  removeElement = () => {
-    this.#element = null;
   }
 }
